requirements.txt
arxiv==2.1.0
requests==2.31.0
pyyaml==6.0.1
loguru==0.7.2
pandas==2.2.0
pymupdf==1.23.26
tqdm==4.66.2
sentence-transformers==2.5.1
faiss-cpu==1.8.0
numpy==1.26.4
ollama==0.1.7
streamlit==1.32.0
networkx==3.2.1
pyvis==0.3.2
